backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
fastapi==0.136.1
frozenlist==1.8.0
h11==0.16.0
idna==3.13
multidict==6.7.1
propcache==0.4.1
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
starlette==1.0.0
twilio==9.10.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
yarl==1.23.0
